364:Nombres d'éléments affiché incorrect et récapitulatif de l'envoi non affiché

See merge request mc/francetransfert/francetransfert-download-api!79

+++ src/main/java/fr/gouv/culture/francetransfert/application/resources/model/DirectoryRepresentation.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DirectoryRepresentation extends DataRepresentation {
-    private int totalSize;
+    private long totalSize;
 }
